refactor(api): move eventLogginJobController to shared
